+++ Source/MindKeeper.Domain/Interfaces/IUserRepository.cs
@@ -6,7 +6,7 @@ namespace MindKeeper.Domain.Interfaces
 {
     public interface IUserRepository
     {
-        public Task<User> Get(int id);
+        public Task<User> Get(long id);
         public Task<User> Get(string mail, bool isNormalizedSearch = false);
         public Task<List<User>> GetAll();
         public Task<User> Create(string mail, string passwordHash);
